Created three new functions
First function was finding the word frequency of a vector of strings, which created and returned an un-ordered map of the frequency distribution. Next function was changing a number to a binary string, which will be used for the next function. The last function implemented is using the map created from the word frequency function, creating a priority queue (pq), then ordering the pq starting at 0 for the highest frequency of a word. Then the number used as a priority is converted to a binary string from the num to binary string function. Last, the word and the new binary representation of it is added to a map and that map is returned. This will be used to save to the two files.